% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\docType{data}
\name{abundance}
\alias{abundance}
\title{Abundance measurements}
\format{A data frame with 5832 rows and 7 variables}
\usage{
data(abundance)
}
\description{
A dataset containing abundances for each plant species, where each species was sampled at its developmental peak.
}
\details{
\itemize{      
  \item year: year
  \item month: month of sampling
  \item day: day of sampling
  \item plot: plot
  \item subplot: subplot code
  \item species: plant species
  \item individuals: number of individuals
  }
}
\note{
For details, see Lanuza et al. 2018 Ecology Letters.
}
\keyword{datasets}
