% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BH_2.R
\name{BH_2}
\alias{BH_2}
\title{Title Beverton-Holt fecundity, second model}
\usage{
BH_2(par, param.list, log.fitness, focal.comp.matrix, num.covariates,
  num.competitors, focal.covariates, fixed.terms)
}
\arguments{
\item{par}{vector of variable length, with the following order: first, lambda of focal sp; 
second alpha, single interaction coefficient; 
last, sigma value. If any element is not to be optimized, it must not be present in this vector, but rather in the "fixed.terms" list}

\item{param.list}{string listing parameters to optimize. Possible elements are \code{lambda}, \code{lambda.cov}, \code{alpha}, \code{alpha.cov}.}

\item{log.fitness}{log of fitness value}

\item{focal.comp.matrix}{dataframe with as many rows as observations, and one column for each competitor sp. 
Values of the dataframe are number of competitors of each sp per observation.}

\item{num.covariates}{not used in BH_2}

\item{num.competitors}{not used in BH_2}

\item{focal.covariates}{not used in BH_2.}

\item{fixed.terms}{list with elements \code{lambda}, \code{lambda.cov}, \code{alpha}, \code{alpha.cov}. It contains parameters not to be optimized.
Each element of the list must be of its appropriate length. Note that adding an element in "param.list" will force the function
to look for it in \code{par}, and will not consider it here. In this model, \code{lambda.cov} and \code{alpha.cov} are not considered.}
}
\value{
log-likelihood value
}
\description{
These functions return the negative log-likelihood of the data
given the model and parameters. BH_2 is $F_i = \lambda_i/(1+\alpha*N_{j})$
}
