% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BH_abundance_1.R
\name{BH_abundance_1}
\alias{BH_abundance_1}
\title{Project abundance of individuals according to the Beverton-Holt second model}
\usage{
BH_abundance_1(sp.par, init.abund, cov.values, alpha.matrix,
  lambda.cov.matrix, alpha.cov.matrix, return.seeds = TRUE)
}
\arguments{
\item{sp.par}{dataframe with species in rows, and the following columns:
lambda: fecundity term
germ.rate: seed germination rate
survival.rate: annual survival of ungerminated seed}

\item{init.abund}{number of individuals at time t}

\item{cov.values}{Not used in BH_abundances_1}

\item{alpha.matrix}{Not used in BH_1}

\item{lambda.cov.matrix}{Not used in BH_abundances_1}

\item{alpha.cov.matrix}{Not used in BH_abundances_1}

\item{return.seeds}{boolean flag, whether the prediction should return 
number of seeds (i.e. $N_{i,t+1}$ eq. 1 of Lanuza et al. 2018), or number of
adult individuals, (i.e. $N_{i,t+1} * g$ )}
}
\value{
1d vector with number of individuals of each species at time t+1
}
\description{
Project abundance of individuals according to the Beverton-Holt second model
}
