% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BH_abundance_4.R
\name{BH_abundance_4}
\alias{BH_abundance_4}
\title{Project abundance of individuals according to the Beverton-Holt fourth model}
\usage{
BH_abundance_4(sp.par, init.abund, cov.values, alpha.matrix,
  lambda.cov.matrix, alpha.cov.matrix)
}
\arguments{
\item{sp.par}{dataframe with species in rows, and the following columns:
lambda: fecundity term
germ.rate: seed germination rate
survival.rate: annual survival of ungerminated seed}

\item{init.abund}{number of individuals at time t}

\item{cov.values}{1d vector with values of each covariate}

\item{alpha.matrix}{competition matrix}

\item{lambda.cov.matrix}{matrix of  num.sp x num.cov
giving the effect of each covariate over the fecundity (lambda) of each species}

\item{alpha.cov.matrix}{list of dimension number of covariates. 
Each component of the list is a single value,
giving the effect of the covariate in question over the interaction matrix.}

\item{return.seeds}{boolean flag, whether the prediction should return 
number of seeds (i.e. $N_{i,t+1}$ eq. 1 of Lanuza et al. 2018), or number of
adult individuals, (i.e. $N_{i,t+1} * g$ )}
}
\value{
1d vector with number of individuals of each species at time t+1
}
\description{
Project abundance of individuals according to the Beverton-Holt fourth model
}
