% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BevertonHolt_abundance_models.R
\name{BH_abundance_5}
\alias{BH_abundance_5}
\title{Title}
\usage{
BH_abundance_5(sp.par, init.abund, cov.values, alpha.matrix,
  lambda.cov.matrix, alpha.cov.matrix)
}
\arguments{
\item{sp.par}{dataframe with species in rows, and the following columns:
lambda: fecundity term
germ.rate: seed germination rate
survival.rate: annual survival of ungerminated seed}

\item{init.abund}{number of individuals at time t}

\item{cov.values}{1d vector with values of each covariate}

\item{alpha.matrix}{competition matrix}

\item{lambda.cov.matrix}{matrix of dimension number of species x number of covariates
giving the effect of each covariate over the fecundity (lambda) of each species}

\item{alpha.cov.matrix}{list of dimension number of covariates. 
Each component of the list is a matrix of dimension number of species x number of species,
giving the effect of the covariate in question over the interaction matrix.}
}
\value{
1d vector with number of individuals of each species at time t+1
}
\description{
Title
}
