% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BH_niche_overlap.R
\name{BH_niche_overlap}
\alias{BH_niche_overlap}
\title{niche overlap function for Beverton-Holt (BH) models}
\usage{
BH_niche_overlap(pair_matrix)
}
\arguments{
\item{pair_matrix}{pairwise interaction matrix}
}
\value{
numeric value
}
\description{
niche overlap function for Beverton-Holt (BH) models
}
\examples{
BH_niche_overlap(matrix(c(.1,.2,.03,.3)))
}
