% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/BH_project_alpha_pairwise_lambdacov_none_alphacov_none.R
\name{BH_project_alpha_pairwise_lambdacov_none_alphacov_none}
\alias{BH_project_alpha_pairwise_lambdacov_none_alphacov_none}
\title{Beverton-Holt model for projecting abundances,
with specific alpha values and no covariate effects}
\usage{
BH_project_alpha_pairwise_lambdacov_none_alphacov_none(lambda, alpha_intra,
  alpha_inter, lambda_cov, alpha_cov, abundance, covariates)
}
\arguments{
\item{lambda}{numeric lambda value.}

\item{alpha_intra}{included for compatibility, not used in this model.}

\item{alpha_inter}{single numeric value.}

\item{lambda_cov}{included for compatibility, not used in this model.}

\item{alpha_cov}{included for compatibility, not used in this model.}

\item{abundance}{named numeric vector of abundances in the previous timestep.}

\item{covariates}{included for compatibility, not used in this model.}
}
\value{
numeric abundance projected one timestep
}
\description{
Beverton-Holt model for projecting abundances,
with specific alpha values and no covariate effects
}
