% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompetitiveAbility.R
\name{CompetitiveAbility}
\alias{CompetitiveAbility}
\title{Competitive ability of a species}
\usage{
CompetitiveAbility(lambda, germ.rate = NULL, survival.rate = NULL,
  competitive.response)
}
\arguments{
\item{lambda}{per capita fecundity of the species in the absence of competition}

\item{germ.rate}{germination rate of the species}

\item{survival.rate}{annual survival of ungerminated seed in the soil}

\item{competitive.response}{parameter reflecting the species' sensitivity to competition}
}
\value{
competitive ability of the species
}
\description{
Calculates the competitive ability of a species following eq. S2 of Godoy et al. (2014). If germ.rate and survival.rate are provided, 
it calculates nu according to eq. 4 of Godoy et al. (2014), for annual plants. Otherwise, if only lambda is provided, nu = lambda
}
