% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxr-package.R
\docType{package}
\name{cxr-package}
\alias{cxr}
\alias{cxr-package}
\title{cxr: A Toolbox for Modelling Species Coexistence in R}
\description{
Recent developments in modern coexistence theory have advanced our understanding on how species are able to persist and co-occur with other species at varying abundances. However, applying this mathematical framework to empirical data is still challenging, precluding a larger adoption of the theoretical tools developed by empiricists. This package provides a complete toolbox for modelling interaction effects between species, and calculate fitness and niche differences. The functions are flexible, may accept covariates, and different fitting algorithms can be used. A full description of the underlying methods is available in García-Callejas, D., Godoy, O., and Bartomeus, I. (2020) \doi{10.1111/2041-210X.13443}. Furthermore, the package provides a series of functions to calculate dynamics for stage-structured populations across sites.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RadicalCommEcol/cxr}
  \item Report bugs at \url{https://github.com/RadicalCommEcol/cxr/issues}
}

}
\author{
\strong{Maintainer}: David Garcia-Callejas \email{david.garcia.callejas@gmail.com} (\href{https://orcid.org/0000-0001-6982-476X}{ORCID})

Authors:
\itemize{
  \item Ignasi Bartomeus \email{nacho.bartomeus@gmail.com} (\href{https://orcid.org/0000-0001-7893-4389}{ORCID})
  \item Oscar Godoy \email{ogodoy.re@gmail.com} (\href{https://orcid.org/0000-0003-4988-6626}{ORCID})
}

Other contributors:
\itemize{
  \item Maxime Lancelot [contributor]
  \item Maria Paniw [contributor]
}

}
\keyword{internal}
