% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxr_optimize.R
\name{cxr_optimize}
\alias{cxr_optimize}
\title{General optimization for dynamic models}
\usage{
cxr_optimize(fitness.model, optim.method, param.list, log.fitness,
  init.lambda = NULL, lower.lambda = 1, upper.lambda = 1e+05,
  init.sigma = NULL, lower.sigma = 1e-10, upper.sigma = 1e+05,
  init.alpha = 1e-04, lower.alpha = 0, upper.alpha = 1e+05,
  init.lambda.cov = 0.001, lower.lambda.cov = 1e-04,
  upper.lambda.cov = 1e+05, init.alpha.cov = 0.001,
  lower.alpha.cov = 1e-04, upper.alpha.cov = 1e+05, focal.comp.matrix,
  focal.covariates = NULL, generate.errors = FALSE,
  bootstrap.samples = 0, verbose = FALSE)
}
\arguments{
\item{fitness.model}{function giving the population dynamics model. Any functional form is allowed, but the model must be constrained
to free parameters \code{lambda} (fecundity of each sp in absence of competition), \code{alpha} (interaction coefficients),
\code{lambda.cov} (effect of covariates on lambda), \code{alpha.cov} (effect of covariates on alpha)}

\item{optim.method}{optimization method to use. One of the following: "optim_NM","optim_L-BFGS-B","nloptr_CRS2_LM", 
"nloptr_ISRES","nloptr_DIRECT_L_RAND","GenSA","hydroPSO","DEoptimR".}

\item{param.list}{string vector giving the parameters that are to be optimized for the fitness model.}

\item{log.fitness}{1d vector, log of the fitness metric for every observation}

\item{init.lambda}{1d vector, initial value of lambda}

\item{lower.lambda}{lower bound for lambda}

\item{upper.lambda}{upper bound for lambda}

\item{init.sigma}{initial value for sigma (standard deviation)}

\item{lower.sigma}{lower bound for sigma}

\item{upper.sigma}{upper bound for sigma}

\item{init.alpha}{initial value for the alpha vector/matrix}

\item{lower.alpha}{lower bound for alpha}

\item{upper.alpha}{upper bound for alpha}

\item{init.lambda.cov}{initial value for the lambda.cov matrix. Discarded if no covariates are given.}

\item{lower.lambda.cov}{lower bound for lambda.cov}

\item{upper.lambda.cov}{upper bound for lambda.cov}

\item{init.alpha.cov}{initial value for the alpha.cov matrix. Discarded if no covariates are given.}

\item{lower.alpha.cov}{lower bound for alpha.cov}

\item{upper.alpha.cov}{upper bound for alpha.cov}

\item{focal.comp.matrix}{matrix with observations in rows and neighbours in columns. Each cell is the number of neighbours
of a given species in a given observation.}

\item{focal.covariates}{optional matrix with observations in rows and covariates in columns. Each cell is the value of a covariate
in a given observation.}

\item{generate.errors}{boolean, whether to compute bootstrap errors for the fitted parameters. Note that, depending on 
the data, model, and optimization method, this may be computationally expensive.}

\item{bootstrap.samples}{how many bootstrap samples to compute.}
}
\value{
list with the fitted parameters, and the loglikelihood of the fit. If a parameter is taken as a constant, the list will return 
the original value given.
}
\description{
Wrapper for optimization procedures. It accepts a model, defined as a function, and a series of parameters, and returns
the optimal value for the parameters given a fitness metric and an optimization method. Optionally, bootstrap confidence intervals
can also be computed.
}
