% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxr_pm_bootstrap.R
\name{cxr_pm_bootstrap}
\alias{cxr_pm_bootstrap}
\title{Standard error estimates for model parameters}
\usage{
cxr_pm_bootstrap(fitness_model, optimization_method, data, covariates,
  init_par, lower_bounds, upper_bounds, fixed_parameters,
  bootstrap_samples)
}
\arguments{
\item{fitness_model}{function returning a single value to minimize, given a set of parameters and a fitness metric}

\item{optimization_method}{numerical optimization method}

\item{data}{dataframe with observations in rows and two sets of columns:
\itemize{
\item fitness: fitness metric for the focal individual
\item neighbours: columns with user-defined names with number of neighbours for each group
}}

\item{covariates}{optional matrix with observations in rows and covariates in columns. Each cell is the value of a covariate
in a given observation.}

\item{init_par}{1d vector of initial parameters}

\item{lower_bounds}{1d vector of lower bounds}

\item{upper_bounds}{1d vector of upper bounds}

\item{fixed_parameters}{optional list specifying values of fixed parameters, with components "lambda","alpha","lambda_cov", and "alpha_cov".}

\item{bootstrap_samples}{how many bootstrap samples to compute.}
}
\value{
1d vector, the standard error of each parameter in init.par
}
\description{
Computes bootstrap standard errors for a given population dynamics model.
}
