% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxr_retrieve_params.R
\name{cxr_retrieve_params}
\alias{cxr_retrieve_params}
\title{Retrieve parameters from the vector returned by the optimization procedures}
\usage{
cxr_retrieve_params(optim.params, param.list, alpha.length,
  alpha.cov.length, num.competitors, num.covariates)
}
\arguments{
\item{optim.params}{1d vector, the result of an optimization method}

\item{param.list}{character vector, which parameters are present. Possible elements are "lambda", "lambda.cov", "alpha", "alpha.cov".}

\item{alpha.length}{if alpha is to be retrieved, its length}

\item{alpha.cov.length}{if alpha.cov is to be retrieved, its length}

\item{num.competitors}{how many competitor species}

\item{num.covariates}{how many covariates}
}
\value{
list with elements "lambda", "alpha", "lambda.cov", "alpha.cov", "sigma". If one of these elements is not present, returns NULL.
}
\description{
Retrieve parameters from the vector returned by the optimization procedures
}
