% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densities_to_df.R
\name{densities_to_df}
\alias{densities_to_df}
\title{Converts a densities list to a tidy dataframe}
\usage{
densities_to_df(densities)
}
\arguments{
\item{densities}{list, species (optionally x year) with each element holding 
a sites x stages matrix. This function assumes three life stages.}
}
\value{
dataframe with columns species-stage-site(-year)-density
}
\description{
Converts a densities list to a tidy dataframe
}
