% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ER_optimize.R
\name{ER_optimize}
\alias{ER_optimize}
\title{Estimate competition effects and responses for a set of species}
\usage{
ER_optimize(lambda.vector, e.vector, r.vector, lambda.cov = NULL,
  e.cov = NULL, r.cov = NULL, sigma, lambda.lower.bound = 0,
  lambda.upper.bound = 1000, e.lower.bound, e.upper.bound, r.lower.bound,
  r.upper.bound, lambda.cov.lower.bound = NULL,
  lambda.cov.upper.bound = NULL, e.cov.lower.bound = NULL,
  e.cov.upper.bound = NULL, r.cov.lower.bound = NULL,
  r.cov.upper.bound = NULL, sigma.lower.bound, sigma.upper.bound,
  effect.response.model, optim.method, sp.data, covariates = NULL,
  optimize.lambda = FALSE, generate.errors = FALSE,
  bootstrap.samples = 0)
}
\arguments{
\item{lambda.vector}{1d vector of lambda estimates/initial values (depending on whether lambda values are optimized or not)}

\item{e.vector}{1d vector of competitive effect initial values}

\item{r.vector}{1d vector of competitive response initial values}

\item{lambda.cov}{numeric matrix of num.sp x num.covariates, effect of every covariate on species' lambda. 
Discarded if covariates are not passed.}

\item{e.cov}{numeric matrix of num.sp x num.covariates, effect of every covariate on species' competitive effect.
Discarded if covariates are not passed.}

\item{r.cov}{numeric matrix of num.sp x num.covariates, effect of every covariate on species' competitive response. 
Discarded if covariates are not passed.}

\item{sigma}{initial value for variation estimate.}

\item{lambda.lower.bound}{lower bounds for lambda, in case it is optimized. Same dimensions as lambda.vector}

\item{lambda.upper.bound}{upper bounds for lambda, in case it is optimized. Same dimensions as lambda.vector}

\item{e.lower.bound}{lower bounds for competitive effects. Same dimensions as e.vector}

\item{e.upper.bound}{upper bounds for competitive effects. Same dimensions as e.vector}

\item{r.lower.bound}{lower bounds for competitive responses. Same dimensions as r.vector}

\item{r.upper.bound}{upper bounds for competitive responses. Same dimensions as r.vector}

\item{lambda.cov.lower.bound}{lower bounds for covariate effects on lambda. Same dimensions as lambda.cov. 
Discarded if covariates are not passed.}

\item{lambda.cov.upper.bound}{upper bounds for covariate effects on lambda. Same dimensions as lambda.cov. 
Discarded if covariates are not passed.}

\item{e.cov.lower.bound}{lower bounds for covariate effects on e. Same dimensions as e.cov. 
Discarded if covariates are not passed.}

\item{e.cov.upper.bound}{upper bounds for covariate effects on e. Same dimensions as e.cov. 
Discarded if covariates are not passed.}

\item{r.cov.lower.bound}{lower bounds for covariate effects on r. Same dimensions as r.cov. 
Discarded if covariates are not passed.}

\item{r.cov.upper.bound}{upper bounds for covariate effects on r. Same dimensions as r.cov. 
Discarded if covariates are not passed.}

\item{sigma.lower.bound}{lower bound for sigma.}

\item{sigma.upper.bound}{upper bound for sigma.}

\item{effect.response.model}{function returning a value to optimize over, e.g. maximum likelihood}

\item{optim.method}{optimization method to use. One of the following: "optim_NM","optim_L-BFGS-B","nloptr_CRS2_LM", 
"nloptr_ISRES","nloptr_DIRECT_L_RAND","GenSA","hydroPSO","DEoptimR".}

\item{sp.data}{dataframe with all the necessary information in long format. It should have the following columns:
- site: character ID
- focal: character ID of the focal species. Any number of focal species is allowed, but the number of focal species
must match the number of initial parameters (one lambda, e, and r per species).
- fitness: numeric, a fitness metric
- competitor: character, ID of a competitor for that observation. The set of competitors must be, for now, the same
as the set of focal species.
- number: number of neighbouring/competitor individuals from the associated species. Observations without competitors of a given species
must be explicit, i.e. setting number to zero.}

\item{covariates}{optional matrix/dataframe with as many rows as observations and covariates in columns.}

\item{optimize.lambda}{boolean, whether we want to optimize lambda values or not}

\item{generate.errors}{boolean, whether to compute bootstrap errors for the fitted parameters. Note that, depending on 
the data and optimization method, this may be computationally expensive.}

\item{bootstrap.samples}{how many bootstrap samples to compute.}
}
\value{
list with estimated species values for e, r, lambda (optional), and if covariates are given, the effects of covariates on lambda, r, and e.
}
\description{
This function is similar in spirit to cxr_optimize, in that it optimizes a set of parameters
via maximum likelihood.
}
