% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_demography_matrix.R
\name{fill_demography_matrix}
\alias{fill_demography_matrix}
\title{Fill the vec-permutation demography matrix}
\usage{
fill_demography_matrix(focal.sp, vpm, transition_matrices)
}
\arguments{
\item{focal.sp}{integer, focal species.}

\item{vpm}{data structure holding all vector-permutation matrices; see 
`vec_permutation_matrices`. If not in an appropriate format, it is likely
to fail without warning.}

\item{transition_matrices}{nested list species x sites, in which each element
holds a 3x3 transition matrix. If not in that format, it is likely to fail
without warning.}
}
\value{
vec-permutation demography matrix for a given species across sites.
}
\description{
Fill for a given species, across all sites.
}
