% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dispersal_matrix.R
\name{fill_dispersal_matrix}
\alias{fill_dispersal_matrix}
\title{Fill the vec-permutation dispersal matrix}
\usage{
fill_dispersal_matrix(
  focal.sp,
  num.sites,
  param,
  vpm,
  env = NULL,
  current.densities
)
}
\arguments{
\item{focal.sp}{integer, focal species}

\item{num.sites}{integer, how many sites}

\item{param}{param nested list,see `build_param` function}

\item{vpm}{data structure holding all vector-permutation matrices; see 
`vec_permutation_matrices`}

\item{env}{optional numeric, environmental forcing for a given timestep}

\item{current.densities}{list of length sp, each element is a matrix sites*stages}
}
\value{
dispersal matrix, stages*sites
}
\description{
Fill for a given species, all sites
}
