% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_transition_matrix.R
\name{fill_transition_matrix}
\alias{fill_transition_matrix}
\title{Fill a transition matrix}
\usage{
fill_transition_matrix(focal.sp, site, param, env = NULL, current.densities)
}
\arguments{
\item{focal.sp}{integer, species}

\item{site}{integer, site}

\item{param}{param structure (see `build_param` function)}

\item{env}{optional numeric, environmental forcing for a given timestep}

\item{current.densities}{list of length sp, each element is a matrix site*stages}
}
\value{
3x3 transition matrix
}
\description{
Calculates the elements of a site-specific transition matrix for a given sp.
Note that here, and through all functions, we fix three life stages.
Also note that `param` and `env` must match, as for the `vital_rate` function.
}
