% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_ratio.R
\name{fitness_ratio}
\alias{fitness_ratio}
\title{Fitness ratio among two or more species}
\usage{
fitness_ratio(
  effect_response_fit = NULL,
  fitness_sp1 = NULL,
  fitness_sp2 = NULL
)
}
\arguments{
\item{effect_response_fit}{cxr_er_fit object}

\item{fitness_sp1}{numeric value representing the fitness (a.k.a. competitive ability) of the first taxa}

\item{fitness_sp2}{numeric value representing the fitness (a.k.a. competitive ability) of the second taxa}
}
\value{
either a matrix with fitness ratios for all pairs of fitted species, or a single numeric value.
The matrix elements represent the ratios of species in columns over species in rows, and conversely, 
the numeric value represents the ratio of sp1 over sp2.
}
\description{
Fitness ratio among two or more species
}
\examples{
fitness_ratio(fitness_sp1 = 0.6, fitness_sp2 = 0.3)
}
