% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateTestData.R
\name{GenerateTestData}
\alias{GenerateTestData}
\title{Generate simulated competition data}
\usage{
GenerateTestData(focal.sp = 1, num.sp = 2, num.cov = 2,
  num.obs = 10, fitness.model = 1, focal.lambda, alpha,
  alpha.cov = NULL, lambda.cov = NULL)
}
\arguments{
\item{focal.sp}{number of focal species}

\item{num.sp}{total number of species, including focal ones}

\item{num.cov}{number of covariates}

\item{num.obs}{number of observations/sites}

\item{fitness.model}{model to generate data from, from 1 to 5 in increasing levels of complexity.
The generating models are equivalent to BH_1-BH_5}

\item{focal.lambda}{1d vector with lambdas of the focal sp}

\item{alpha}{interaction matrix, num.sp x num.sp}

\item{alpha.cov}{list of dimension num.cov. 
Each component of the list is a matrix of different dimensions depending on \code{fitness.model}.
If fitness.model is 4, each component should be a single value, giving the effect of each covariate over every interaction; 
if fitness.model is 5, each component should be a matrix num.sp x num.sp,
giving the effect of the covariate in question over each element of the interaction matrix.}

\item{lambda.cov}{matrix of num.sp x num.cov
giving the effect of each covariate over the fecundity (lambda) of each species}
}
\value{
dataset with a fitness metric calculated for each focal species and observation, according to the fitness model selected
}
\description{
Generate simulated competition data
}
