% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_example_coefs.R
\docType{data}
\name{glm_example_coefs}
\alias{glm_example_coefs}
\title{Generalized linear model coefficients}
\format{
A named numerical matrix of 8 rows and 4 columns
}
\usage{
data(glm_example_coefs)
}
\description{
A table with coefficients from a GLM to serve as an example for importing
into the data structure of the metapopulation model.
}
\keyword{datasets}
