% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitParams.R
\name{InitParams}
\alias{InitParams}
\title{Join parameters in a 1d vector}
\usage{
InitParams(init.lambda = NULL, init.sigma = 0, init.alpha = NULL,
  init.lambda.cov = NULL, init.alpha.cov = NULL, lower.lambda = 1,
  upper.lambda = 1e+05, lower.sigma = 1e-10, upper.sigma = 1e+05,
  lower.alpha = 0, upper.alpha = 1e+05, lower.lambda.cov = 1e-04,
  upper.lambda.cov = 1e+05, lower.alpha.cov = 1e-04,
  upper.alpha.cov = 1e+05, num.competitors, num.covariates)
}
\arguments{
\item{init.lambda}{numeric, lambda}

\item{init.sigma}{numeric, sigma}

\item{init.alpha}{1d vector, interaction coefficients over the species}

\item{init.lambda.cov}{1d vector, initial values for lambda.cov}

\item{init.alpha.cov}{1d vector, initial values for alpha.cov}

\item{lower.lambda}{lower bound for lambda}

\item{upper.lambda}{upper bound for lambda}

\item{lower.sigma}{lower bound for sigma}

\item{upper.sigma}{upper bound for sigma}

\item{lower.alpha}{lower bound for alpha}

\item{upper.alpha}{upper bound for alpha}

\item{lower.lambda.cov}{lower bound for lambda.cov}

\item{upper.lambda.cov}{upper bound for lambda.cov}

\item{lower.alpha.cov}{lower bound for alpha.cov}

\item{upper.alpha.cov}{upper bound for alpha.cov}

\item{num.competitors}{number of competitors}

\item{num.covariates}{number of covariates}
}
\value{
list with three 1d vectors, ready for passing to the optim methods, consistent with the functions BH_1-5
}
\description{
Generate a 1d vector from a series of parameters in a certain order. It also returns the same vector
for lower and upper bounds. This function is intended to work with
parameters for a single species (i.e. a single lambda value, etc). Note that lambda.cov and alpha.cov must be consistent
with num.covariates.
}
