% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LV_niche_overlap.R
\name{LV_niche_overlap}
\alias{LV_niche_overlap}
\title{niche overlap function for Lotka-Volterra (LV) models}
\usage{
LV_niche_overlap(pair_matrix)
}
\arguments{
\item{pair_matrix}{pairwise interaction matrix}
}
\value{
numeric value
}
\description{
niche overlap function for Lotka-Volterra (LV) models
}
\examples{
LV_niche_overlap(matrix(c(.1,.2,.03,.3),nrow = 2))
}
