% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/LV_project_alpha_pairwise_lambdacov_global_alphacov_global.R
\name{LV_project_alpha_pairwise_lambdacov_global_alphacov_global}
\alias{LV_project_alpha_pairwise_lambdacov_global_alphacov_global}
\title{Lotka-Volterra model for projecting abundances,
with specific alpha values and global covariate effects on alpha and lambda}
\usage{
LV_project_alpha_pairwise_lambdacov_global_alphacov_global(lambda,
  alpha_intra, alpha_inter, lambda_cov, alpha_cov, abundance, covariates)
}
\arguments{
\item{lambda}{numeric lambda value.}

\item{alpha_intra}{single numeric value.}

\item{alpha_inter}{numeric vector with interspecific alpha values.}

\item{lambda_cov}{numeric vector with effects of covariates over lambda.}

\item{alpha_cov}{named list of numeric values 
with effects of each covariate over alpha.}

\item{abundance}{named numeric vector of abundances in the previous timestep.}

\item{covariates}{matrix with observations in rows and covariates in columns. Each cell is the value of a covariate
in a given observation.}
}
\value{
numeric abundance projected one timestep
}
\description{
Lotka-Volterra model for projecting abundances,
with specific alpha values and global covariate effects on alpha and lambda
}
