% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/LW_er_lambdacov_none_effectcov_none_responsecov_none.R
\name{LW_er_lambdacov_none_effectcov_none_responsecov_none}
\alias{LW_er_lambdacov_none_effectcov_none_responsecov_none}
\title{Effect response Law-Watkinson model without covariate effects}
\usage{
LW_er_lambdacov_none_effectcov_none_responsecov_none(par, fitness, target,
  density, covariates, fixed_parameters)
}
\arguments{
\item{par}{1d vector with initial parameters in the order: 
lambda,effect,response,sigma.}

\item{fitness}{1d vector with fitness observations.}

\item{target}{matrix with species in rows, observations in columns. Value is 1 if
a species is focal for a given observation, 0 otherwise.}

\item{density}{matrix with species in rows, observations in columns. Value is 
density of each sp as neighbour for each observation.}

\item{covariates}{included for compatibility, not used in this model.}

\item{fixed_parameters}{optional list specifying values of fixed parameters, 
with components "lambda","effect","response".}
}
\value{
log-likelihood value
}
\description{
Note that, as e and r are not pair-specific, all species parameters are fit in the same function.
}
