% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LW_niche_overlap.R
\name{LW_niche_overlap}
\alias{LW_niche_overlap}
\title{niche overlap function for Lqw-Watkinson (LW) models}
\usage{
LW_niche_overlap(pair_matrix)
}
\arguments{
\item{pair_matrix}{pairwise interaction matrix}
}
\value{
numeric value
}
\description{
niche overlap function for Lqw-Watkinson (LW) models
}
\examples{
LW_niche_overlap(matrix(c(.1,.2,.03,.3),nrow = 2))
}
