% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/LW_project_alpha_pairwise_lambdacov_none_alphacov_none.R
\name{LW_project_alpha_pairwise_lambdacov_none_alphacov_none}
\alias{LW_project_alpha_pairwise_lambdacov_none_alphacov_none}
\title{Law-Watkinson model for projecting abundances,
with specific alpha values and no covariate effects}
\usage{
LW_project_alpha_pairwise_lambdacov_none_alphacov_none(
  lambda,
  alpha_intra,
  alpha_inter,
  lambda_cov,
  alpha_cov,
  abundance,
  covariates
)
}
\arguments{
\item{lambda}{numeric lambda value.}

\item{alpha_intra}{included for compatibility, not used in this model.}

\item{alpha_inter}{single numeric value.}

\item{lambda_cov}{included for compatibility, not used in this model.}

\item{alpha_cov}{included for compatibility, not used in this model.}

\item{abundance}{named numeric vector of abundances in the previous timestep.}

\item{covariates}{included for compatibility, not used in this model.}
}
\value{
numeric abundance projected one timestep
}
\description{
Law-Watkinson model for projecting abundances,
with specific alpha values and no covariate effects
}
