% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_ER_lambda.R
\name{model_ER_lambda}
\alias{model_ER_lambda}
\title{Estimation of competitive effects and responses}
\usage{
model_ER_lambda(init.par, target_all, density_all, log.fitness,
  covariates = NULL)
}
\arguments{
\item{init.par}{1d vector of initial parameters: lambda values, r values, e values, and single sigma term.
If covariates are given, this vector must include also lambda.cov, response.cov, and effect.cov terms, 
after the e values and before the sigma term.}

\item{target_all}{matrix giving which species is calculated with which values. See \code{er_optim}}

\item{density_all}{matrix giving the densities of each species at each observation. See \code{er_optim}}

\item{log.fitness}{log of the fitness metric}

\item{covariates}{if present, it is a dataframe/matrix with as many rows as observationes, and one column for each covariate.
Values are covariate values for every observation.}
}
\value{
single numeric value giving the sum of negative log-likelihoods
}
\description{
Calculates the log-likelihood of a Beverton-Holt model parameterized with given values
with respect to a fitness metric. The function for calculating fecundity given 
effect and response values is taken from Godoy et al. (2014). 
Note that, as e is not pair-specific, all species parameters are fit in the same function.
In this version, lambda values are also fit.
}
