% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neigh_list.R
\docType{data}
\name{neigh_list}
\alias{neigh_list}
\title{neighbours and fitness observations}
\format{A list with 16 elements, each of which a dataframe
of variable number of rows and 18 columns}
\usage{
data(neigh_list)
}
\description{
A dataset containing fitness and neighbours 
for plant individuals of 16 species. 
The dataset is a named list with 16 elements,
each of which is a dataframe with the following columns:
}
\details{
\itemize{          
  \item obs_ID: unique identifier for each observation
  \item fitness: number of viable seeds of the focal individual
  \item 16 columns indicating the number of neighbours from each plant sp.
  in a radius of 7.5 cm from the focal individual
  }
}
\note{
For details, see Lanuza et al. 2018 Ecology Letters.
}
\keyword{datasets}
