% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/pm_BH_alpha_global_lambdacov_none_alphacov_none.R
\name{pm_BH_alpha_global_lambdacov_none_alphacov_none}
\alias{pm_BH_alpha_global_lambdacov_none_alphacov_none}
\title{Beverton-Holt model with a global alpha and no covariate effects}
\usage{
pm_BH_alpha_global_lambdacov_none_alphacov_none(par, fitness, neigh_matrix,
  covariates, fixed_parameters)
}
\arguments{
\item{par}{1d vector of initial parameters: lambda, alpha, and sigma}

\item{fitness}{1d vector of fitness observations, in log scale}

\item{neigh_matrix}{matrix with number of neighbours (each neighbour a column) for each observation (in rows)}

\item{covariates}{included for compatibility, not used in this model}

\item{fixed_parameters}{optional list specifying values of fixed parameters, 
with components "lambda","alpha".}
}
\value{
log-likelihood value
}
\description{
Beverton-Holt model with a global alpha and no covariate effects
}
