% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/pm_BH_alpha_pairwise_lambdacov_global_alphacov_global.R
\name{pm_BH_alpha_pairwise_lambdacov_global_alphacov_global}
\alias{pm_BH_alpha_pairwise_lambdacov_global_alphacov_global}
\title{Beverton-Holt model with pairwise alphas and global covariate effects on lambda and alpha}
\usage{
pm_BH_alpha_pairwise_lambdacov_global_alphacov_global(par, fitness,
  neigh_matrix, covariates, fixed_parameters)
}
\arguments{
\item{par}{1d vector of initial parameters: lambda, lambda_cov, alpha, alpha_cov, and sigma}

\item{fitness}{1d vector of fitness observations, in log scale}

\item{neigh_matrix}{matrix with number of neighbours (each neighbour a column) for each observation (in rows)}

\item{covariates}{optional matrix with observations in rows and covariates in columns. Each cell is the value of a covariate
in a given observation}

\item{fixed_parameters}{optional list specifying values of fixed parameters, 
with components "lambda","alpha","lambda_cov","alpha_cov".}
}
\value{
log-likelihood value
}
\description{
Beverton-Holt model with pairwise alphas and global covariate effects on lambda and alpha
}
