% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm_LV_alpha_none_lambdacov_none_alphacov_none.R
\name{pm_LV_alpha_none_lambdacov_none_alphacov_none}
\alias{pm_LV_alpha_none_lambdacov_none_alphacov_none}
\title{Lotka-Volterra model with no alphas and no covariate effects}
\usage{
pm_LV_alpha_none_lambdacov_none_alphacov_none(par, fitness,
  neigh_intra_matrix = NULL, neigh_inter_matrix, covariates,
  fixed_parameters)
}
\arguments{
\item{par}{1d vector of initial parameters: lambda and sigma}

\item{fitness}{1d vector of fitness observations, in log scale}

\item{neigh_intra_matrix}{included for compatibility, not used in this model.}

\item{neigh_inter_matrix}{included for compatibility, not used in this model.}

\item{covariates}{included for compatibility, not used in this model}

\item{fixed_parameters}{included for compatibility, not used in this model}
}
\value{
log-likelihood value
}
\description{
This model, in all families, is simply given by lambda.
}
