% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/pm_RK_alpha_global_lambdacov_none_alphacov_none.R
\name{pm_RK_alpha_global_lambdacov_none_alphacov_none}
\alias{pm_RK_alpha_global_lambdacov_none_alphacov_none}
\title{Ricker model with a global alpha and no covariate effects}
\usage{
pm_RK_alpha_global_lambdacov_none_alphacov_none(par, fitness,
  neigh_intra_matrix = NULL, neigh_inter_matrix, covariates,
  fixed_parameters)
}
\arguments{
\item{par}{1d vector of initial parameters: lambda, alpha, and sigma.}

\item{fitness}{1d vector of fitness observations, in log scale.}

\item{neigh_intra_matrix}{included for compatibility, not used in this model.}

\item{neigh_inter_matrix}{matrix of arbitrary columns, number of neighbours for each observation.
As in this model there is a single alpha argument, do not distinguish neighbour identity}

\item{covariates}{included for compatibility, not used in this model.}

\item{fixed_parameters}{optional list specifying values of fixed parameters, 
with components "lambda","alpha_inter".}
}
\value{
log-likelihood value
}
\description{
Ricker model with a global alpha and no covariate effects
}
