% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/RK_er_lambdacov_global_effectcov_global_responsecov_global.R
\name{RK_er_lambdacov_global_effectcov_global_responsecov_global}
\alias{RK_er_lambdacov_global_effectcov_global_responsecov_global}
\title{Effect response Beverton-Holt model with covariate effects on lambda, effect, and response}
\usage{
RK_er_lambdacov_global_effectcov_global_responsecov_global(
  par,
  fitness,
  target,
  density,
  covariates,
  fixed_parameters
)
}
\arguments{
\item{par}{1d vector with initial parameters in the order: 
lambda,lambda_cov,effect,effect_cov,response,response_cov,sigma}

\item{fitness}{1d vector with fitness observations}

\item{target}{matrix with species in rows, observations in columns. Value is 1 if
a species is focal for a given observation, 0 otherwise.}

\item{density}{matrix with species in rows, observations in columns. Value is 
density of each sp as neighbour for each observation.}

\item{covariates}{numeric dataframe or matrix with observations 
in rows and covariates in columns. Each cell is the value of a covariate
in a given observation}

\item{fixed_parameters}{optional list specifying values of fixed parameters, 
with components "lambda","lambda_cov","effect","effect_cov",
"response","response_cov".}
}
\value{
log-likelihood value
}
\description{
Note that, as e and r are not pair-specific, all species parameters are fit in the same function.
}
