% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RK_niche_overlap.R
\name{RK_niche_overlap}
\alias{RK_niche_overlap}
\title{niche overlap function for Ricker (RK) models}
\usage{
RK_niche_overlap(pair_matrix)
}
\arguments{
\item{pair_matrix}{pairwise interaction matrix}
}
\value{
numeric value
}
\description{
niche overlap function for Ricker (RK) models
}
\examples{
RK_niche_overlap(matrix(c(.1,.2,.03,.3)))
}
