% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinity.R
\docType{data}
\name{salinity}
\alias{salinity}
\title{Salinity measurements}
\format{A data frame with 324 rows and 5 variables}
\usage{
data(salinity)
}
\description{
A dataset containing the integrated salinity over the season for each plant species.
}
\details{
\itemize{
  \item plot: plot number
  \item subplot: subplot code
  \item year: year of measurement
  \item precip: anual precipitation
  \item sum_salinity: summatory of the salinity values integrated througth the species lifespan. 
  }
}
\note{
For details, see Lanuza et al. 2017 Ecology Letters.
}
\keyword{datasets}
