% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEbootstrap.R
\name{SEbootstrap}
\alias{SEbootstrap}
\title{Standard error estimates for model parameters}
\usage{
SEbootstrap(fitness.model, optim.method, param.list, fixed.terms,
  log.fitness, init.par, lower.bounds, upper.bounds, focal.comp.matrix,
  focal.covariates, nsamples)
}
\arguments{
\item{fitness.model}{function returning a loglikelihood value given a set of parameters and a fitness metric}

\item{optim.method}{optimization method to use. One of the following: "optim_NM","optim_L-BFGS-B","nloptr_CRS2_LM", 
"nloptr_ISRES","nloptr_DIRECT_L_RAND","GenSA","hydroPSO","DEoptimR".}

\item{param.list}{string vector giving the parameters that are to be optimized for the fitness model.}

\item{fixed.terms}{string vector giving the parameters that are NOT optimized for the fitness model.}

\item{log.fitness}{1d vector, log of the fitness metric for every observation}

\item{init.par}{1d vector of initial parameters}

\item{lower.bounds}{1d vector of lower bounds}

\item{upper.bounds}{1d vector of upper bounds}

\item{focal.comp.matrix}{matrix with observations in rows and neighbours in columns. Each cell is the number of neighbours
of a given species in a given observation.}

\item{focal.covariates}{optional matrix with observations in rows and covariates in columns. Each cell is the value of a covariate
in a given observation.}

\item{nsamples}{how many bootstrap samples to compute.}
}
\value{
1d vector, the standard error of each parameter in init.par
}
\description{
Computes bootstrap standard errors for a given population dynamics model.
}
