% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_sampling.R
\docType{data}
\name{spatial_sampling}
\alias{spatial_sampling}
\title{spatial arrangement of the observations}
\format{
A list with 16 elements, each of which a dataframe
of variable number of rows and 18 columns
}
\usage{
data(spatial_sampling)
}
\description{
A dataset giving the spatial arrangement of observations.
The dataset is a list of 16 elements following the structure of
'neigh_list'. Each list component is a dataframe
with columns:
}
\details{
\itemize{          
  \item obs_ID: unique identifier for each observation
  \item plot: one of 9 plots of 8.5 x 8.5 m
  \item subplot: one of 36 subplots of 1x1 m within each plot
  }
}
\note{
For details, see Lanuza et al. 2018 Ecology Letters.
}
\keyword{datasets}
