% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_rates.R
\docType{data}
\name{species_rates}
\alias{species_rates}
\title{Species germination and survival rates}
\format{A data frame with 16 rows and 4 variables}
\usage{
data(species_rates)
}
\description{
A dataset containing germination and survival rates for 16 plant species. 
It includes columns with the scientific names and their associated codes.
}
\details{
\itemize{      
  \item species: binomial name
  \item code: four-letter code used in other datasets
  \item germination: germination rate
  \item seed.survival: annual survival of ungerminated seed in the soil
  }
}
\note{
For details, see Lanuza et al. 2018 Ecology Letters.
}
\keyword{datasets}
