% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.cxr_pm_multifit}
\alias{summary.cxr_pm_multifit}
\title{CXR summary method for multispecies fits}
\usage{
\method{summary}{cxr_pm_multifit}(object, ...)
}
\arguments{
\item{object}{a \code{cxr_pm_multifit} object, from the function with the same name}

\item{...}{other arguments, not used}
}
\value{
console output
}
\description{
CXR summary method for multispecies fits
}
