% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\docType{data}
\name{abundance}
\alias{abundance}
\title{Abundance measurements}
\format{A data frame with 5184 rows and 4 variables}
\usage{
data(abundance)
}
\description{
A dataset containing abundances for each plant species, 
where each species was sampled at its developmental peak.
}
\details{
\itemize{      
  \item plot: one of 9 plots of the study area
  \item subplot: one of 36 1x1 m subplots of each plot
  \item species: plant species
  \item individuals: number of individuals
  }
}
\note{
For details, see Lanuza et al. 2018 Ecology Letters.
}
\keyword{datasets}
