% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BH_1.R
\name{BH_1}
\alias{BH_1}
\title{Title Beverton-Holt fecundity, first model}
\usage{
BH_1(par, param.list, log.fitness, focal.comp.matrix, num.covariates,
  num.competitors, focal.covariates, fixed.terms)
}
\arguments{
\item{par}{vector containing lambda of focal sp and sigma value}

\item{param.list}{not used in BH_1}

\item{log.fitness}{log of fitness value}

\item{focal.comp.matrix}{not used in BH_1}

\item{num.covariates}{not used in BH_1}

\item{num.competitors}{not used in BH_1}

\item{focal.covariates}{not used in BH_1}

\item{fixed.terms}{not used in BH_1}
}
\value{
log-likelihood value
}
\description{
These functions return the negative log-likelihood of the data
given the model and parameters. BH_1 is $F_i = \lambda_i$
}
