% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BevertonHolt_abundance_models.R
\name{BH_abundance_3}
\alias{BH_abundance_3}
\title{Title}
\usage{
BH_abundance_3(sp.par, init.abund, cov.values, alpha.matrix,
  lambda.cov.matrix, alpha.cov.matrix)
}
\arguments{
\item{sp.par}{dataframe with species in rows, and the following columns:
lambda: fecundity term
germ.rate: seed germination rate
survival.rate: annual survival of ungerminated seed}

\item{init.abund}{number of individuals at time t}

\item{cov.values}{Not used in BH_abundances_3}

\item{alpha.matrix}{competition matrix}

\item{lambda.cov.matrix}{Not used in BH_abundances_3}

\item{alpha.cov.matrix}{Not used in BH_abundances_3}
}
\value{
1d vector with number of individuals of each species at time t+1
}
\description{
Title
}
\examples{
#no example
}
