% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition.R
\docType{data}
\name{competition}
\alias{competition}
\title{Competition measurments}
\format{A data frame with 58176 rows and 10 variables}
\usage{
data(competition)
}
\description{
A dataset containing fitness and neighbours for each plant species.
}
\details{
\itemize{          
  \item year: year
  \item month: month
  \item day: day
  \item plot: plot
  \item subplot: subplot code
  \item focal: focal plant species identity
  \item fruit: total fruits produced
  \item seed: total seeds produced
  \item competitor: competitor identity
  \item number: number of competitors
  }
}
\note{
For details, see Lanuza et al. 2017 Ecology Letters.
}
\keyword{datasets}
