% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateTestData.R
\name{GenerateTestData}
\alias{GenerateTestData}
\title{Generate simulated competition data}
\usage{
GenerateTestData(focal.sp = 1, num.sp = 2, num.cov = 2,
  num.obs = 10, fitness.model = 1, focal.lambda, alpha.matrix,
  alpha.cov, lambda.cov)
}
\arguments{
\item{focal.sp}{number of focal species}

\item{num.sp}{total number of species, including focal ones}

\item{num.cov}{number of covariates}

\item{num.obs}{number of observations/sites}

\item{fitness.model}{model to generate data from, from 1 to 5 in increasing levels of complexity.
The generating models are equivalent to BH_1-BH_5}

\item{focal.lambda}{1d vector with lambdas of the focal sp}

\item{alpha.matrix}{num.sp x num.sp interaction matrix}

\item{alpha.cov}{list containing num.sp matrices of size num.sp * num.cov, giving the effect of a given covariate -column- on the interaction alpha i -list index- ,j -matrix row-}

\item{lambda.cov}{list containing num.sp vectors giving the effect of each covariate on the lambda of species i}
}
\value{
dataset with a fitness metric calculated for each focal species and observation, according to the fitness model selected
}
\description{
Generate simulated competition data
}
