% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictAbundances.R
\name{PredictAbundances}
\alias{PredictAbundances}
\title{Project abundances according to specified models and parameters}
\usage{
PredictAbundances(par, timesteps, abundance.model, return.seeds = TRUE)
}
\arguments{
\item{par}{list with the following components: 
1 - dataframe sp.par, with the parameters to be passed to the predictive model
2 - dataframe initial.values, with fields "site","species","abundance"
3 - covariates, either 0 if there are no covariates, or a dataframe with fields "site","timestep","covariate","value"
4 - list other.par, other parameters to \code{abundance model}, such as lambda.cov.matrix, alpha.cov.matrix.
species should be a unique identifier, character or numeric. idem for site and covariate.}

\item{timesteps}{number of timesteps to project}

\item{abundance.model}{a function that accepts parameters from sp.par, a set of initial abundances, and optionally other parameters.
The function returns the projected abundances at t+1}

\item{return.seeds}{boolean flag, whether the prediction should return 
number of seeds (i.e. $N_{i,t+1}$ eq. 1 of Lanuza et al. 2018), or number of
adult individuals, (i.e. $N_{i,t+1} * g$ )}
}
\value{
dataframe with fields "timestep","site","sp","abundance", giving the expected abundance for each species, timestep, and site.
}
\description{
Project abundances according to specified models and parameters
}
